+++ src/MovieApp/MovieApp.Logic/Models/UserMoviePreference.cs
@@ -3,7 +3,7 @@ namespace MovieApp.Logic.Models
     public class UserMoviePreference
     {
         public int Id { get; set; }
-        public double Score { get; set; }
+        public decimal Score { get; set; }
         public DateTime LastModified { get; set; }
         public int? ChangeFromPreviousValue { get; set; }
